requirements.txt
python-dotenv==1.1.0
openai==1.76.0
neo4j==5.28.1
gradio==5.27.0
pyvis==0.3.2
